requirements.txt
asgiref==3.3.1
Django==3.1.6
django-cors-headers==3.7.0
django-summernote==0.8.11.6
djangorestframework==3.12.2
Pillow==8.2.0
psycopg2==2.8.6
python-dotenv==0.17.0
pytz==2021.1
sqlparse==0.4.1
